jax3d/projects/mobilenerf/requirements.txt
numpy>=1.16.4
jax>=0.2.6
jaxlib>=0.1.69
flax>=0.2.2
opencv-python>=4.4.0
Pillow>=7.2.0
matplotlib>=3.3.4
trimesh>=4.0.0
